api_erp_app/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-talisman>=1.1.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "joblib>=1.4.2",
    "numpy>=2.2.4",
    "openai>=1.72.0",
    "pika>=1.3.2",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "redis>=5.2.1",
    "scikit-learn>=1.6.1",
]
